third_party/blink/web_tests/fast/dom/shadow/offset-parent-does-not-leak-ua-shadow-expected.txt

CONSOLE WARNING: line 33: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
offsetParent should not leak nodes in user agent Shadow DOM.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS child.offsetParent is shadow.positionedElement
PASS child.offsetParent is container
PASS child.offsetParent is container
PASS child.offsetParent is container
PASS successfullyParsed is true

TEST COMPLETE
